% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_na.R
\name{mean_na}
\alias{mean_na}
\title{Calculate Mean Ignoring NAs and Non-Numeric Values}
\usage{
mean_na(x)
}
\arguments{
\item{x}{A vector containing numeric and/or non-numeric values.}
}
\value{
The mean of the numeric values in the vector.
}
\description{
Calculate Mean Ignoring NAs and Non-Numeric Values
}
\examples{
mean_na(c(1, 2, "a", NA, 4, 5, "b", 5))
}
