% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode_na.R
\name{mode_na}
\alias{mode_na}
\title{Calculate Mode Ignoring NAs and Non-Numeric Values}
\usage{
mode_na(x)
}
\arguments{
\item{x}{A vector containing numeric and/or non-numeric values.}
}
\value{
The mode of the numeric values in the vector.
}
\description{
Calculate Mode Ignoring NAs and Non-Numeric Values
}
\examples{
mode_na(c(1, 2, "a", NA, 4, 5, "b", 5))
}
