% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stats.R
\name{summary_stats}
\alias{summary_stats}
\title{Summary Statistics Ignoring NAs and Non-Numeric Values}
\usage{
summary_stats(x)
}
\arguments{
\item{x}{A vector containing numeric and/or non-numeric values.}
}
\value{
A list with the mean, median, standard deviation, minimum, and maximum of the numeric values in the vector.
}
\description{
Summary Statistics Ignoring NAs and Non-Numeric Values
}
\examples{
summary_stats(c(1, 2, "a", NA, 4, 5, "b", 5))
}
