% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniform_sex.R
\name{uniform_sex}
\alias{uniform_sex}
\title{Normalize Sex/Gender Values}
\usage{
uniform_sex(x)
}
\arguments{
\item{x}{A character vector containing sex/gender values.}
}
\value{
A character vector with normalized sex/gender values.
}
\description{
Normalize Sex/Gender Values
}
\examples{
uniform_sex(c("M", "D", "F", "W", "m", "d", "f", "w"))
}
